:lipstick: api: rename usecases into libUsecases
